+++ VERMAXION/Configuration.cs
@@ -21,6 +21,7 @@ public class Configuration : IPluginConfiguration
     public string DtrIconDisabled { get; set; } = "\uE03D";
     public bool KrangleEnabled { get; set; } = false;
     public bool AutoRestoreRetainerCheckingAfterWork { get; set; } = true;
+    public bool EnableCharacterSelectStallRecovery { get; set; } = true;
     public float LeftPanelWidth { get; set; } = 240f;
     public CharacterListSortMode CharacterListSortMode { get; set; } = CharacterListSortMode.Name;
     public List<string> PostProcessTaskOrder { get; set; } = VERMAXION.PostProcessTaskOrder.DefaultOrder.ToList();